Refactor pagination module to reduce its size
